fix:    vueファイルを分解して、tsに変換
