+++ package.json
@@ -23,6 +23,7 @@
     "not ie <= 8"
   ],
   "devDependencies": {
+    "@types/marked": "^0.4.0",
     "babel-core": "^6.26.0",
     "babel-loader": "^7.1.2",
     "babel-preset-env": "^1.6.0",
