+++ src/app/components/Editor/editor.scss
@@ -1,0 +1,40 @@
+.editor-wrapper {
+  display: flex;
+}
+.memo-list-wrapper {
+  width: 19%;
+  float: left;
+  border-top: 1px solid #000;
+}
+.memo-list {
+  padding: 10px;
+  box-sizing: border-box;
+  text-align: left;
+  border-bottom: 1px solid #000;
+  &:nth-clild(even) {
+    background-color: #ccc;
+  }
+  &[data-selected="true"] {
+    background-color: #ccf;
+  }
+}
+.memo-title {
+  height: 1.5em;
+  margin: 0;
+  white-space: nowrap;
+  overflow: hidden
+}
+.add-memo-btn {
+  margin-top: 20px;
+}
+.delete-memo-btn {
+  margin: 10px;
+}
+.markdown {
+  width: 40%;
+  height: 500px;
+}
+.preview {
+  width: 40%;
+  height: 500px;
+}
